Removing many redundant API tests.
