+++ tests/Alchemy/Tests/Phrasea/Controller/Api/ApiJSONPTest.php
@@ -11,30 +11,6 @@ use Symfony\Component\HttpFoundation\Response;
  */
 class ApiJSONPTest extends ApiTestCase
 {
-    protected function evaluateResponseBadRequest(Response $response)
-    {
-        $this->assertEquals('UTF-8', $response->getCharset(), 'Test charset response');
-        $this->assertEquals(200, $response->getStatusCode(), 'Test status code 400 ' . $response->getContent());
-    }
-
-    protected function evaluateResponseForbidden(Response $response)
-    {
-        $this->assertEquals('UTF-8', $response->getCharset(), 'Test charset response');
-        $this->assertEquals(200, $response->getStatusCode(), 'Test status code 403 ' . $response->getContent());
-    }
-
-    protected function evaluateResponseNotFound(Response $response)
-    {
-        $this->assertEquals('UTF-8', $response->getCharset(), 'Test charset response');
-        $this->assertEquals(200, $response->getStatusCode(), 'Test status code 404 ' . $response->getContent());
-    }
-
-    protected function evaluateResponseMethodNotAllowed(Response $response)
-    {
-        $this->assertEquals('UTF-8', $response->getCharset(), 'Test charset response');
-        $this->assertEquals(200, $response->getStatusCode(), 'Test status code 405 ' . $response->getContent());
-    }
-
     protected function getParameters(array $parameters = [])
     {
         $parameters['callback'] = 'jsFunction';
